# apps/rpg-cyborgs/openapi/rpg.json
{
  "openapi": "3.0.1",
  "info": {
      "title": "Rpg API v1",
      "version": "1.0"
  },
  "paths": {
      "/api/rpg/{system}/{archetype}/{id}": {
          "get": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  },
                  {
                      "name": "archetype",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  },
                  {
                      "name": "id",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_String"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_String"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_String"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/api/rpg/{system}/describe": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeProp"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeProp"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeProp"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeProp"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_PropDescription"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_PropDescription"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_PropDescription"
                              }
                          }
                      }
                  },
                  "400": {
                      "description": "Bad Request"
                  }
              }
          }
      },
      "/api/rpg/{system}/entities": {
          "get": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "type": "array",
                                  "items": {
                                      "$ref": "#/components/schemas/Server.RpgContent"
                                  }
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "type": "array",
                                  "items": {
                                      "$ref": "#/components/schemas/Server.RpgContent"
                                  }
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "type": "array",
                                  "items": {
                                      "$ref": "#/components/schemas/Server.RpgContent"
                                  }
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/modset": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ModSet"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ModSet"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ModSet"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_ModSet"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Boolean"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Boolean"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_Boolean"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/modset/describe": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeModSet"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeModSet"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeModSet"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeModSet"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_ModSetDescription"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_ModSetDescription"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_ModSetDescription"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/state": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_SetState"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_SetState"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_SetState"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_SetState"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_String"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_String"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_String"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/state/describe": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeState"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeState"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeState"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_DescribeState"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_ModSetDescription"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_ModSetDescription"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_ModSetDescription"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/api/rpg/{system}/time": {
          "post": {
              "tags": [
                  "v1"
              ],
              "parameters": [
                  {
                      "name": "system",
                      "in": "path",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_PointInTime"
                          }
                      },
                      "text/json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_PointInTime"
                          }
                      },
                      "application/*+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_PointInTime"
                          }
                      },
                      "application/json-patch+json": {
                          "schema": {
                              "$ref": "#/components/schemas/Server.RpgRequest_PointInTime"
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_PointInTime"
                              }
                          },
                          "text/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_PointInTime"
                              }
                          },
                          "text/plain": {
                              "schema": {
                                  "$ref": "#/components/schemas/Server.RpgResponse_PointInTime"
                              }
                          }
                      }
                  }
              }
          }
      }
  },
  "components": {
      "schemas": {
...
          "Mods.ModSets.StateModSet": {
              "required": [
                  "instanceType",
                  "stateName"
              ],
              "type": "object",
              "allOf": [
                  {
                      "$ref": "#/components/schemas/Mods.ModSet"
                  }
              ],
              "properties": {
                  "stateName": {
                      "type": "string"
                  },
                  "instanceType": {
                      "$ref": "#/components/schemas/States.StateInstanceType"
                  }
              },
              "additionalProperties": false
          },
...
          "Props.PropDescription": {
              "required": [
                  "baseValue",
                  "entityArchetype",
                  "entityId",
                  "entityName",
                  "mods",
                  "prop",
                  "rootEntityArchetype",
                  "rootEntityId",
                  "rootEntityName",
                  "rootProp",
                  "value"
              ],
              "type": "object",
              "properties": {
                  "rootEntityId": {
                      "type": "string"
                  },
                  "rootEntityName": {
                      "type": "string"
                  },
                  "rootEntityArchetype": {
                      "type": "string"
                  },
                  "rootProp": {
                      "type": "string"
                  },
                  "entityId": {
                      "type": "string"
                  },
                  "entityName": {
                      "type": "string"
                  },
                  "entityArchetype": {
                      "type": "string"
                  },
                  "prop": {
                      "type": "string"
                  },
                  "value": {
                      "$ref": "#/components/schemas/Values.Dice"
                  },
                  "baseValue": {
                      "$ref": "#/components/schemas/Values.Dice"
                  },
                  "mods": {
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/Props.ModDescription"
                      }
                  }
              },
              "additionalProperties": false
          },
...
          "RpgGraphState": {
              "required": [
                  "entities",
                  "time"
              ],
              "type": "object",
              "properties": {
                  "entities": {
                      "type": "array",
                      "items": {
                          "oneOf": [
                              {
                                  "$ref": "#/components/schemas/Cyborgs.Armour"
                              },
                              {
                                  "$ref": "#/components/schemas/Cyborgs.BodyPart"
                              },
                              {
                                  "$ref": "#/components/schemas/Cyborgs.MeleeWeapon"
                              },
                              {
                                  "$ref": "#/components/schemas/Cyborgs.PlayerCharacter"
                              },
                              {
                                  "$ref": "#/components/schemas/Cyborgs.RangedWeapon"
                              },
                              {
                                  "$ref": "#/components/schemas/Cyborgs.Components.PropValue"
                              },
                              {
                                  "$ref": "#/components/schemas/RpgContainer"
                              },
                              {
                                  "$ref": "#/components/schemas/Actions2.Action2"
                              },
                              {
                                  "$ref": "#/components/schemas/Actions2.Activity2"
                              }
                          ]
                      }
                  },
                  "contextId": {
                      "type": "string",
                      "nullable": true
                  },
                  "time": {
                      "$ref": "#/components/schemas/Time.Temporal"
                  }
              },
              "additionalProperties": false
          },
...
          "Server.Ops.DescribeProp": {
              "required": [
                  "entityId",
                  "prop"
              ],
              "type": "object",
              "properties": {
                  "entityId": {
                      "type": "string"
                  },
                  "prop": {
                      "type": "string"
                  }
              },
              "additionalProperties": false
          },
...
          "Server.RpgContent": {
              "required": [
                  "archetype",
                  "key",
                  "name",
                  "system"
              ],
              "type": "object",
              "properties": {
                  "key": {
                      "type": "string",
                      "format": "uuid"
                  },
                  "system": {
                      "type": "string"
                  },
                  "archetype": {
                      "type": "string"
                  },
                  "name": {
                      "type": "string"
                  }
              },
              "additionalProperties": false
          },
...
          "Server.RpgRequest_DescribeModSet": {
              "required": [
                  "graphState",
                  "op"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "op": {
                      "$ref": "#/components/schemas/Server.Ops.DescribeModSet"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgRequest_DescribeProp": {
              "required": [
                  "graphState",
                  "op"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "op": {
                      "$ref": "#/components/schemas/Server.Ops.DescribeProp"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgRequest_DescribeState": {
              "required": [
                  "graphState",
                  "op"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "op": {
                      "$ref": "#/components/schemas/Server.Ops.DescribeState"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgRequest_ModSet": {
              "required": [
                  "graphState"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "op": {
                      "oneOf": [
                          {
                              "$ref": "#/components/schemas/Mods.ModSet"
                          },
                          {
                              "$ref": "#/components/schemas/Mods.ModSets.StateModSet"
                          },
                          {
                              "$ref": "#/components/schemas/Mods.ModSets.SyncedModSet"
                          },
                          {
                              "$ref": "#/components/schemas/Mods.ModSets.TimedModSet"
                          },
                          {
                              "$ref": "#/components/schemas/Mods.ModSets.TurnModSet"
                          }
                      ],
                      "nullable": true
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgRequest_PointInTime": {
              "required": [
                  "graphState",
                  "op"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "op": {
                      "$ref": "#/components/schemas/Time.PointInTime"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgRequest_SetState": {
              "required": [
                  "graphState",
                  "op"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "op": {
                      "$ref": "#/components/schemas/Server.Ops.SetState"
                  }
              },
              "additionalProperties": false
          },
...
          "Server.RpgResponse_Boolean": {
              "required": [
                  "data",
                  "graphState"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "data": {
                      "type": "boolean"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgResponse_ModSetDescription": {
              "required": [
                  "data",
                  "graphState"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "data": {
                      "$ref": "#/components/schemas/Mods.ModSetDescription"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgResponse_PointInTime": {
              "required": [
                  "data",
                  "graphState"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "data": {
                      "$ref": "#/components/schemas/Time.PointInTime"
                  }
              },
              "additionalProperties": false
          },
          "Server.RpgResponse_PropDescription": {
              "required": [
                  "data",
                  "graphState"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "data": {
                      "$ref": "#/components/schemas/Props.PropDescription"
                  }
              },
              "additionalProperties": false
          },
...
          "Server.RpgResponse_String": {
              "required": [
                  "graphState"
              ],
              "type": "object",
              "properties": {
                  "graphState": {
                      "$ref": "#/components/schemas/RpgGraphState"
                  },
                  "data": {
                      "type": "string",
                      "nullable": true
                  }
              },
              "additionalProperties": false
          },
...
      },
      "securitySchemes": {
          "Backoffice User": {
              "type": "oauth2",
              "description": "Umbraco Authentication",
              "flows": {
                  "authorizationCode": {
                      "authorizationUrl": "/umbraco/management/api/v1/security/back-office/authorize",
                      "tokenUrl": "/umbraco/management/api/v1/security/back-office/token",
                      "scopes": {}
                  }
              }
          }
      }
  }
}